more clean up / remove notebooks
